We now have basic combat set up
I mapped an attack button and an item button to the keyboard (1 to
attack, 2 to use an item). Added background I drew earlier. You will
need to have the debug log open to see the damage dealt. I'll try to
clean things up a little bit and maybe add on screen text.