Moved Closer to Combat Functionality
UserControllable and Monster now have Id's,
Monster creation logic now exists, where it picks monster(s) based on
the level. One to 4 monsters can be chosen, and since some Monsters are
tougher than others and we don't want to overwhelm the player, Monsters
have a difficultWithinLevel variable, ranging from 1 to 4. The sum
difficultyWithinLevel of a fight can never exceed 4 (so you can get
attacked by a monster w/ 1 difficultyInLevel and 3 difficultyInLevel,
(sums to 4) but not a monster with 2 difficultyInLEvel and 3
difficultyInLevel (sums to 5). Implemented Attack ability's cast. You
click on the ability, then if there's only 1 target it auto-attacks,
otherwise it waits for you to select target.  To do this it sets a
delegate called "pipeIntoFunc", that an ability can set. If it's set,
all UI button's click functions go to the ability's functtion it set.
pipeIntoFunc takes a string with two arguments - 1. the type of thing
that was clicked (ability bar, userCotnrollable, monster, etc) and the
index of that thing. Implemented clicking on a userControllable image
loads their ability bar. In battlescript, there is now a monster[]
called monsters, and userControllable called activeUserControllable
(which is what userControllable currenlty has access to the ability
menu). I also created a genericBenchmarkMonster class, and tested out
loading a prefab into unity (the exact syntax proved tricky and took me
hours to figure out).