Added enemy animations
I added an enterance and idle animation for demon skull. The animations
are currently used in a separate scene that I created to test them. The
images for the spirtes and the animation files are already in the
assets. I plan on trying to bring the animations into the main game this
weekend