Worked on Proof-of-concept for damage-text.
UserControllables now have damage-text and status-effect-text. We have a
minor issue that we need to resolve as a team regarding how to color the
text, whether it should have a background panel (to make the color stand
out), and whether we want it as a crisp "UI text" or less crisp "Text
mesh." The latter has the advantage of being able to be displayed on top
of enemies... do we want that?