Put together the beginning code for shop / inventory
InventoryItem is now the superclass for Item. It contains invObject
which is the gameObject that the represents the item in the
shop/inventory scene. There is a script called ShopInventoryScript that
handles the user clicking on "Weapon", "Item", or "Loot" for both the
shop / inventory. Basicaly, it's changes which ScrollView is currently
active.